feat: Refactor conversation streaming and implement suggestions streaming endpoint
- Updated APIHandler to change the URL for suggestions streaming.
- Removed the old streamConversation method from conversationController.
- Added a new streamSuggestions method in suggestionController for handling streaming suggestions.
- Updated conversationRoutes to remove the old streaming endpoint and added the new suggestions streaming endpoint in suggestionRoutes.
