+++ server/controllers/suggestionController.js
@@ -5,6 +5,8 @@ import crypto from 'crypto';
 import { chatCompletion as mistralChatCompletion } from '../services/mistralService.js';
 import { chatCompletion as openAIChatCompletion } from '../services/openaiService.js';
 import { buildAssistantSuggestionPrompt } from '../services/promptBuilder.js';
+import { fetchConversation } from '../controllers/conversationController.js';
+import { streamChatCompletion as mistralStreamChatCompletion } from '../services/mistralService.js';
 
 
 const TIMEOUT = 30000; // Increased to 30 seconds
@@ -313,4 +315,46 @@ export const generateSuggestionsWithFallback = async (req, res) => {
         });
     }
 };
+
+// Stream conversation as server-sent events (SSE)
+export const streamSuggestions = async (req, res) => {
+    const { cid } = req.params;
+    // Set SSE headers
+    res.set({
+      'Content-Type': 'text/event-stream',
+      'Cache-Control': 'no-cache',
+      Connection: 'keep-alive'
+    });
+    res.flushHeaders();
+    try {
+      // Fetch prior conversation + build prompt
+      const memory = await fetchConversation(cid);
+      const messages = buildAssistantSuggestionPrompt(memory);
+  
+      // Call Mistral with streaming
+      let stream;
+      try {
+        stream = await mistralStreamChatCompletion(messages);
+      } catch (streamErr) {
+        console.error('Mistral streaming error', streamErr);
+        res.write(`data: ERROR ${streamErr.message}\n\n`);
+        return res.end();
+      }
+  
+      // Pipe Mistral's SSE directly to client
+      stream.on('data', chunk => {
+        res.write(chunk);
+      });
+      stream.on('end', () => {
+        res.end();
+      });
+      stream.on('error', err => {
+        console.error('Stream error', err);
+        res.end();
+      });
+    } catch (err) {
+      console.error('Streaming failed', err);
+      res.end();
+    }
+  };
